requirements.txt
# Backend Dependencies
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
gunicorn==21.2.0

# Machine Learning
scikit-learn==1.3.0
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
joblib==1.3.2

# Image Processing & OCR
opencv-python==4.8.0.76
pillow==10.0.0
pytesseract==0.3.10

# Text Processing
nltk==3.8.1
textblob==0.17.1
regex==2023.8.8

# Utilities
python-dotenv==1.0.0
requests==2.31.0
